pyproject.toml
[project]
name = "alignment"
version = "1.0.0"
description = "CS 336 Spring 2025 Assignment 5: Alignment"
readme = "README.md"
requires-python = ">=3.11,<3.13"  # Python 3.13 not yet supported for some deps
dependencies = [
    "accelerate>=1.5.2",
    "alpaca-eval",
    "flash-attn==2.7.4.post1",
    "huggingface-hub>=0.20.0",
    "jupyter>=1.1.1",
    "math-verify[antlr4-13-2]>=0.7.0",
    "pylatexenc==2.10",
    "notebook>=7.4.2",
    "pytest>=8.3.5",
    "python-dotenv>=1.0.0",
    "torch",
    "tqdm>=4.67.1",
    "transformers>=4.50.0",
    "typer>=0.15.4",
    "vllm==0.7.2",
    "wandb>=0.19.8",
    "xopen>=2.0.2",
]



[tool.setuptools.packages.find]
include = ["cs336_alignment"]

[tool.uv]
package = true
no-build-isolation-package = ["flash-attn"]


[tool.uv.sources]
alpaca-eval = { git = "https://github.com/nelson-liu/alpaca_eval.git", rev = "forward_kwargs_to_vllm" }

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "WARNING"
addopts = "-s"

[tool.ruff]
line-length = 120

[tool.ruff.lint.extend-per-file-ignores]
# Also ignore `E402` in all `__init__.py` files.
"__init__.py" = ["E402", "F401", "F403", "E501"]
# Ignore bare except errors in grading code
"cs336_alignment/drgrpo_grader.py" = ["E722"]
# Ignore type annotation errors in test files
"tests/**/*.py" = ["E722", "E741", "E711", "UP"]
"scripts/**/*.py" = ["E722", "E741", "E711", "UP"]

[tool.ruff.lint]
extend-select = ["UP"]
ignore = [
    "F722",  # Allow bare except in grading code
    "E741",  # Allow single letter variable names in math code
    "E711",  # Allow == None in some contexts
]

[dependency-groups]
dev = [
    "mypy>=1.16.1",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.0",
    "types-regex>=2024.11.6.20250403",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = "cs336_alignment.drgrpo_grader"
ignore_errors = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "scripts.*"
ignore_errors = true
